move 'composeP' tests to separate file
